+++ plugins/plugin-core-metric-middleware-file-writer/lib/plugins/FileWriterMetricMiddlewarePlugin.ts
@@ -15,23 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * Copyright 2020-2023 Delft University of Technology and SynTest contributors
- *
- * This file is part of SynTest Framework - SynTest Core.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 import { Metric, MetricManager, MiddleWare } from "@syntest/metric";
 import { MetricMiddlewarePlugin } from "@syntest/module";
 import { FileWriterMetricMiddleware } from "../middleware/FileWriterMetricMiddleware";
@@ -63,7 +46,7 @@ export class FileWriterMetricMiddlewarePlugin extends MetricMiddlewarePlugin {
 
     map.set("metrics-directory", {
       alias: [],
-      default: "statistics",
+      default: "metrics",
       description:
         "The path where the csv's should be saved (within the syntest-directory)",
       group: OptionTypes.Storage,